ai_service/requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python
pillow
numpy
python-multipart
distutils